Room prints a list of items and monsters as promised in the iteration.
Inventory has yet to implement iterable interface, so that Room can iterate through
the list of items stored in the Map.

What is next:

Inventory class should implement iterable and consider connecting
sqlite into the database so that we fetch data from database and not use hard coded
elements in Room class.